tools/requirements.txt
svgelements>=1.9
streamlit>=1.35
pillow>=10.0
opencv-python-headless>=4.8
numpy>=1.24
streamlit-image-coordinates>=0.1
boto3>=1.34
cairosvg>=2.7
